Add files via upload
I have added a new section for the spa, in which you can turn off and on the lights, along with turning off and on the jets. I have now added the ability to change/update the low-cost time if needed (However, it still needs some work). I have also added the appropriate watts for each light, and for those lights I have replaced the toggle button with a simple off and on button.