Add files via upload
Was able to add the garage door as a device also added the abiltiy to turn the roller coaster lights, and was able to update the user's input for low-cost times. I also updated the spa control to include an off and on switch like the rest of the toggle controls. I also added a new section that that includes 2 new solar panels devices along with the amount of power the solar panels have recieved within the day.